input replay
