+++ kito/commandframe/commandframe.go
@@ -21,7 +21,6 @@ type CommandFrame struct {
 
 type CommandFrameHistory struct {
 	CommandFrames []CommandFrame
-	ReadOffset    int
 }
 
 func NewCommandFrameHistory() *CommandFrameHistory {
@@ -44,23 +43,35 @@ func (h *CommandFrameHistory) AddCommandFrame(frameNumber int, frameInput input.
 	}
 
 	h.CommandFrames = append(h.CommandFrames, cf)
-
-	// this handles the initial client startup phase where a command frame goes buy without recording a command frame.
-	// this is due to the player not existing yet. there's probably a cleaner way to handle this
-	if len(h.CommandFrames) == 1 {
-		h.ReadOffset = frameNumber
-	}
 }
 
 func (h *CommandFrameHistory) GetCommandFrame(frameNumber int) *CommandFrame {
-	if frameNumber-h.ReadOffset >= len(h.CommandFrames) {
+	if len(h.CommandFrames) == 0 {
 		return nil
 	}
-	return &h.CommandFrames[frameNumber-h.ReadOffset]
+
+	startFrameNumber := h.CommandFrames[0].FrameNumber
+	if frameNumber-startFrameNumber >= len(h.CommandFrames) {
+		return nil
+	}
+
+	return &h.CommandFrames[frameNumber-startFrameNumber]
 }
 
 func (h *CommandFrameHistory) ClearUntilFrameNumber(frameNumber int) {
-	// fmt.Println(frameNumber-h.ReadOffset, len(h.CommandFrames))
-	h.CommandFrames = h.CommandFrames[frameNumber-h.ReadOffset:]
-	h.ReadOffset = frameNumber
+	if len(h.CommandFrames) == 0 {
+		return
+	}
+
+	startFrameNumber := h.CommandFrames[0].FrameNumber
+	h.CommandFrames = h.CommandFrames[frameNumber-startFrameNumber:]
+}
+
+// func (h *CommandFrameHistory) PopUntilFrameNumberInclusive(frameNumber int) {
+// 	startFrameNumber := h.CommandFrames[0].FrameNumber
+// 	h.CommandFrames = h.CommandFrames[frameNumber-startFrameNumber:]
+// }
+
+func (h *CommandFrameHistory) ClearFrames() {
+	h.CommandFrames = []CommandFrame{}
 }
