+++ kito/settings/settings.go
@@ -61,7 +61,7 @@ const (
 
 	// This is potentially overkill to avoiding absolutely no mispredictions on the client.
 	// The drawback of an input buffer is we now add a delay before we process user inputs.
-	MaxInputBufferCommandFrames int = 100 / int(MSPerCommandFrame)
+	MaxInputBufferCommandFrames int = 20 / int(MSPerCommandFrame)
 
 	MaxStateBufferCommandFrames int = 100 / int(MSPerCommandFrame)
